When caching query results, cache all columns, not just those that were requested. Closes #487
