+++ plugin/query_data.go
@@ -97,7 +97,7 @@ type QueryData struct {
 	hydrateCalls []*hydrateCall
 
 	// all the columns that will be returned by this query
-	columns            []*QueryColumn
+	columns            map[string]*QueryColumn
 	concurrencyManager *concurrencyManager
 	rowDataChan        chan *rowData
 	errorChan          chan error
@@ -150,6 +150,7 @@ func newQueryData(connectionCallId string, p *Plugin, queryContext *QueryContext
 		connectionCallId:  connectionCallId,
 		cacheTtl:          executeData.CacheTtl,
 		cacheEnabled:      executeData.CacheEnabled,
+		columns:           make(map[string]*QueryColumn),
 
 		// asyncronously read items using the 'get' or 'list' API
 		// items are streamed on rowDataChan, errors returned on errorChan
@@ -257,29 +258,73 @@ func (d *QueryData) setMatrixItem(matrix []map[string]interface{}) {
 	d.filterMatrixItems()
 }
 
+// build a list of required hydrate function calls which must be executed, based on the columns which have been requested
+// NOTE: 'get' and 'list' calls are hydration functions, but they must be omitted from this list as they are called
+// first. BEFORE the other hydration functions
+// NOTE2: this function also populates the resolvedHydrateName for each column (used to retrieve column values),
+// and the hydrateColumnMap (used to determine which columns to return)
+func (d *QueryData) populateRequiredHydrateCalls() {
+	t := d.Table
+	colsUsed := d.QueryContext.Columns
+	fetchType := d.FetchType
+
+	// what is the name of the fetch call (i.e. the get/list call)
+	fetchFunc := t.getFetchFunc(fetchType)
+	fetchCallName := helpers.GetFunctionName(fetchFunc)
+
+	// initialise hydrateColumnMap
+	d.hydrateColumnMap = make(map[string][]string)
+	requiredCallBuilder := newRequiredHydrateCallBuilder(t, fetchCallName)
+
+	// populate a map keyed by function name to ensure we only store each hydrate function once
+	for _, column := range t.Columns {
+		// see if this column specifies a hydrate function
+
+		var hydrateName string
+		if hydrateFunc := column.Hydrate; hydrateFunc == nil {
+			// so there is NO hydrate call registered for the column
+			// the column is provided by the fetch call
+			// do not add to map of hydrate functions as the fetch call will always be called
+			hydrateFunc = fetchFunc
+			hydrateName = fetchCallName
+		} else {
+			// there is a hydrate call registered
+			hydrateName = helpers.GetFunctionName(hydrateFunc)
+			// if this column was requested in query, add the hydrate call to required calls
+			if helpers.StringSliceContains(colsUsed, column.Name) {
+				requiredCallBuilder.Add(hydrateFunc)
+			}
+		}
+
+		// now update hydrateColumnMap
+		d.hydrateColumnMap[hydrateName] = append(d.hydrateColumnMap[hydrateName], column.Name)
+	}
+	d.hydrateCalls = requiredCallBuilder.Get()
+
+	// now we have all the hydrate calls, build a list of all the columns that will be returned by the hydrate functions.
+	// these will be used for the cache
+
+}
+
 // build list of all columns returned by the fetch call and required hydrate calls
 func (d *QueryData) populateColumns() {
 	// add columns returned by fetch call
 	fetchName := helpers.GetFunctionName(d.Table.getFetchFunc(d.FetchType))
-	d.columns = append(d.columns, d.addColumnsForHydrate(fetchName)...)
+	d.addColumnsForHydrate(fetchName)
 
 	// add columns returned by required hydrate calls
 	for _, h := range d.hydrateCalls {
-		d.columns = append(d.columns, d.addColumnsForHydrate(h.Name)...)
+		d.addColumnsForHydrate(h.Name)
 	}
 }
 
 // get the column returned by the given hydrate call
-func (d *QueryData) addColumnsForHydrate(hydrateName string) []*QueryColumn {
-	var cols []*QueryColumn
+func (d *QueryData) addColumnsForHydrate(hydrateName string) {
 	for _, columnName := range d.hydrateColumnMap[hydrateName] {
 		// get the column from the table
 		column := d.Table.getColumn(columnName)
-		// NOTE: use this to instantiate a QueryColumn
-		// (we cannot use the column directly as we cannot mutate columns as they mayu be shared between tables)
-		cols = append(cols, NewQueryColumn(column, hydrateName))
+		d.columns[columnName] = NewQueryColumn(column, hydrateName)
 	}
-	return cols
 }
 
 // add matrix item into KeyColumnQuals and Quals
